+++ src/Car/PriceValidator.java
@@ -3,7 +3,7 @@ package Car;
 public class PriceValidator implements CarFieldValidator {
     @Override
     public void validate(CarFields carFields) throws FieldValidationException {
-        Integer price = carFields.getPrice();
+        Double price = carFields.getPrice();
         if (price < 0) {
             throw new FieldValidationException("price cannot be negative");
         }
